refactor: 입력값에 대한 예외처리 리팩토링
