+++ docs/README.md
@@ -35,7 +35,7 @@
 <br>
 
 ## 1단계 피드백 체크리스트
-- [ ] 사용하지 않는 메서드 및 import 모두 제거
+- [x] 사용하지 않는 메서드 및 import 모두 제거
 - [x] 자율적인 객체로 리팩토링 (`WinningLotto`)
 - [x] 상수화
     - 로또 두번째 세번째 순위의 맞는 개수인 5를 별도의 상수 처리
