+++ src/main/java/lotto/view/InputView.java
@@ -4,7 +4,9 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
 import java.util.stream.Collectors;
-import lotto.utils.InputValidationUtils;
+import lotto.domain.lotto.Lotto;
+import lotto.domain.lotto.LottoNumber;
+import lotto.exception.LessThanLottoPriceException;
 
 public class InputView {
 
@@ -16,11 +18,17 @@ public class InputView {
         try {
             OutputView.printPurchaseAmountGuideMessage();
             int purchaseAmountValue = Integer.parseInt(scanner.nextLine());
-            InputValidationUtils.validatePurchaseAmount(purchaseAmountValue);
+            validatePurchaseAmount(purchaseAmountValue);
             return purchaseAmountValue;
         } catch (Exception e) {
             OutputView.printExceptionMessage(e);
             return inputPurchaseAmount();
+        }
+    }
+
+    private static void validatePurchaseAmount(int value) {
+        if (value < Lotto.LOTTO_PRICE) {
+            throw new LessThanLottoPriceException();
         }
     }
 
@@ -36,18 +44,23 @@ public class InputView {
             OutputView.printExceptionMessage(e);
             return inputWinningLottoNumbers();
         }
-
     }
 
     public static int inputWinningBonus() {
         try {
             OutputView.printWinningLottoBonusGuideMessage();
             int bonusValue = Integer.parseInt(scanner.nextLine());
-            InputValidationUtils.validateWinningBonus(bonusValue);
+            validateWinningBonus(bonusValue);
             return bonusValue;
         } catch (Exception e) {
             OutputView.printExceptionMessage(e);
             return inputWinningBonus();
+        }
+    }
+
+    public static void validateWinningBonus(int value) {
+        if (value <= LottoNumber.MIN || value > LottoNumber.MAX) {
+            throw new IllegalArgumentException("보너스 번호는 1부터 45사이의 값이어야 합니다.");
         }
     }
 }
